Enable use of host's /nix/store via overlay mount.
While the overlay'd (really, underlay'd) store won't be registered
with the in-chroot nix, it will at least be visible to commands run
inside the chroot (if desired).  This will for instance be useful
during system builds for a remote target, since the system build
process actually seems to want to walk through the local
for some unknown reason.